Further refactoring of typeDefs & resolvers.
