+++ app/graphql/schema/resolvers.js
@@ -1,0 +1,19 @@
+const { get } = require('../../data')
+
+const resolvers = {
+  Query: {
+    customerBusinesses: async (_root, _args, context) => {
+      const response = await get('/organisation/person/3337243/summary?search=', context.crn, context.token)
+      return {
+        crn: context.crn,
+        businesses: response._data?.map(business => ({
+          id: business.id,
+          sbi: business.sbi,
+          name: business.name
+        })) ?? []
+      }
+    }
+  }
+}
+
+module.exports = { resolvers }
