+++ app/graphql/server.js
@@ -1,6 +1,7 @@
 const { ApolloServer } = require('@apollo/server')
-const { get } = require('../data')
-const { typeDefs } = require('./schema')
+// const { get } = require('../data')
+const { typeDefs } = require('./schema/typeDefs')
+const { resolvers } = require('./schema/resolvers')
 
 // const typeDefs = `
 //   type Query {
@@ -19,21 +20,21 @@ const { typeDefs } = require('./schema')
 //   }
 // `
 
-const resolvers = {
-  Query: {
-    customerBusinesses: async (_root, _args, context) => {
-      const response = await get('/organisation/person/3337243/summary?search=', context.crn, context.token)
-      return {
-        crn: context.crn,
-        businesses: response._data?.map(business => ({
-          id: business.id,
-          sbi: business.sbi,
-          name: business.name
-        })) ?? []
-      }
-    }
-  }
-}
+// const resolvers = {
+//   Query: {
+//     customerBusinesses: async (_root, _args, context) => {
+//       const response = await get('/organisation/person/3337243/summary?search=', context.crn, context.token)
+//       return {
+//         crn: context.crn,
+//         businesses: response._data?.map(business => ({
+//           id: business.id,
+//           sbi: business.sbi,
+//           name: business.name
+//         })) ?? []
+//       }
+//     }
+//   }
+// }
 
 const apolloServer = new ApolloServer({
   typeDefs,
